+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Changelog
+
+## Unreleased
+
+### Added
+
+- `cxpo-auto-price`: (new) Adds automatic price calculation
+
 ## 25.4.14
 
 ### Changed