+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `cx-search-bar`: (new) Adds a search bar for materials
+- `cxob-center-on-open`: (new) Centers the order book on open
 - `cxpo-auto-price`: (new) Adds automatic price calculation
 - `macos-antialiased-font`: (new) Applies antialiased smoothing to all fonts on macOS