+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - `cxob-depth-bars`: (new) Adds market depth bars
 - `cxob-hide-section-headers`: (new) Hides "Offers" and "Requests" headers
 - `cxpo-auto-price`: (new) Adds automatic price calculation
+- `cxpo-bigger-buttons`: (new) Makes "Buy" and "Sell" buttons bigger
 - `macos-antialiased-font`: (new) Applies antialiased smoothing to all fonts on macOS
 
 ## 25.4.14