+++ CHANGELOG.md
@@ -4,8 +4,14 @@
 
 ### Added
 
+- `cmds-clickable-commands`: (new) Makes commands clickable
 - `cx-search-bar`: (new) Adds a search bar for materials
+- `cxob-center-on-open`: (new) Centers the order book on open
+- `cxob-depth-bars`: (new) Adds market depth bars
+- `cxob-hide-section-headers`: (new) Hides "Offers" and "Requests" headers
+- `cxob-supply-demand-values`: (new) Adds supply and demand value labels
 - `cxpo-auto-price`: (new) Adds automatic price calculation
+- `cxpo-bigger-buttons`: (new) Makes "Buy" and "Sell" buttons bigger
 - `macos-antialiased-font`: (new) Applies antialiased smoothing to all fonts on macOS
 
 ## 25.4.14